+++ addon/globalPlugins/WordBridge/__init__.py
@@ -29,7 +29,6 @@ from .dialogs import CORRECTOR_CONFIG_FILENAME_DEFAULT, CORRECTOR_CONFIG_FOLDER_
 from .dialogs import LLMSettingsPanel, FeedbackDialog
 from .dictionary.dialog import DictionaryEntryDialog
 from .lib.coseeing import obtain_openai_key
-from .lib.proofreader import Proofreader
 from .lib.typo_corrector import ChineseTypoCorrector, ChineseTypoCorrectorLite
 from .lib.utils import strings_diff
 from .lib.viewHTML import text2template
@@ -218,16 +217,15 @@ class GlobalPlugin(globalPluginHandler.GlobalPlugin):
 				optional_guidance_enable=optional_guidance_enable,
 				customized_words=customized_words,
 			)
-			proofreader = Proofreader(corrector)
 
 			try:
 				batch_mode = not DEBUG_MODE
-				response, _diff_ = proofreader.typo_analyzer(request, batch_mode=batch_mode)
+				response, _diff_ = corrector.correct_text(request, batch_mode=batch_mode)
 			except Exception as e:
 				ui.message(_("Sorry, an error occurred during the program execution, the details are: {e}").format(e=e))
 				log.warning(_("Sorry, an error occurred during the program execution, the details are: {e}").format(e=e))
 				return
-			res = proofreader.get_total_usage()
+			res = corrector.get_total_usage()
 			interaction_id = None
 		else:
 			try:
